requirements.txt
# Flask Web框架
Flask>=2.0.0

# Word文档处理
python-docx>=0.8.11

# PDF处理
PyPDF2>=3.0.0
pdf2docx>=0.5.6

# PDF转图像
pdf2image>=1.16.0

# 图像处理
Pillow>=9.0.0

# 图像转PDF
img2pdf>=0.4.4

# Word转PDF
docx2pdf>=0.1.8
